Rename to JetKit-Flask (#20)
* jetkitmigration

* renaming continued
