+++ .gitignore
@@ -17,6 +17,7 @@ var/
 *.egg
 .mypy_cache/
 
+
 # Serverless directories
 .serverless
 .serverless-wsgi
