+++ jetkit/api/asset/schema.py
