+++ jetkit/api/user/__init__.py
@@ -1,6 +1,6 @@
 from flask_jwt_extended import jwt_required
 from flask_rest_api import Blueprint
-from jb.model.user import CoreUser as User
+from jetkit.model.user import CoreUser as User
 from marshmallow import Schema
 from sqlalchemy.orm import Query
