+++ jetkit/mail/__init__.py
@@ -1,11 +1,11 @@
 """Email clients."""
 
-from jb.mail.constant import MailerImplementation
+from jetkit.mail.constant import MailerImplementation
 
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from jb.mail.base import MailClientBase
+    from jetkit.mail.base import MailClientBase
 
 
 def mail_client(
@@ -15,7 +15,7 @@ def mail_client(
 
     To configure it either pass from_flask or config.
     """
-    from jb.mail.base import MailClientBase
+    from jetkit.mail.base import MailClientBase
 
     return MailClientBase.new_for_impl(impl=impl, from_flask=from_flask, **kwargs)
