+++ setup.py
@@ -3,25 +3,22 @@ from setuptools import setup, find_packages
 with open("README.md") as f:
     long_description = f.read()
 
-NAME = "jb"
-DESCRIPTION = "Common reusable code for python projects."
-VERSION = "5.0.2"
+NAME = "JetKit-Flask"
+DESCRIPTION = "Rapid web application development."
+VERSION = "1.0.0"
 REQUIRES_PYTHON = ">=3.6.0"
 
 setup(
     name=NAME,
     version=VERSION,
     python_requires=REQUIRES_PYTHON,
-    url="https://github.com/jetbridge/backend-core",
+    url="https://github.com/jetbridge/jetkit-flask",
     license="ABRMS",
     author="JetBridge",
     author_email="mischa@jetbridge.com",
     description=DESCRIPTION,
     long_description=long_description,
     long_description_content_type="text/markdown",
-    # py_modules=['jb'],
-    # if you would be using a package instead use packages instead
-    # of py_modules:
     packages=find_packages(exclude=["test", "*.test", "*.test.*", "test.*"]),
     zip_safe=False,
     include_package_data=True,
