+++ tox.ini
@@ -1,15 +1,15 @@
 [flake8]
 ignore = E402,E305,E501,I201,I101,I100,D204,D101
 max-line-length = 160
-exclude = .git,__pycache__,build,dist,.serverless,node_modules,migrations,.venv,jb/model/__init__.py
+exclude = .git,__pycache__,build,dist,.serverless,node_modules,migrations,.venv,jetkit/model/__init__.py
 enable-extensions = pydocstyle,pep8-naming,flake8-debugger,pep8,flake8-docstrings
 
 [pytest]
 # where tests live
-testpaths = jb/test
+testpaths = jetkit/test
 
-addopts = --maxfail=2 --mypy --sqlalchemy-connect-url=postgresql:///jb_core_test --sqlalchemy-manage-db
-mocked-sessions=jb.db.db.session
+addopts = --maxfail=2 --mypy --sqlalchemy-connect-url=postgresql:///jetkit_test --sqlalchemy-manage-db
+mocked-sessions=jetkit.db.db.session
 
 # shut up warnings
 filterwarnings =
